Refactoring to components
